requirements.txt
argon2-cffi==23.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
